refactor: 에디터 footer 리펙토링
