+++ idearly/src/pages/AlgorithmSolvingPage/components/AlgorithmFooter/AlgorithmFooter.styles.tsx
@@ -1,0 +1,20 @@
+import { Button } from "@chakra-ui/react";
+import styled from "@emotion/styled";
+
+export const EditorButton = styled(Button)`
+  font-size: 0.9rem;
+  padding: 0.1rem;
+  width: 3.8rem;
+  height: 2rem;
+  margin-right: 0.5rem;
+`;
+
+export const SubmitButton = styled(Button)`
+  font-size: 0.9rem;
+  padding: 0.1rem;
+  width: 3.8rem;
+  height: 2rem;
+  margin-right: 1rem;
+`;
+
+
