+++ idearly/src/pages/AlgorithmSolvingPage/components/AlgorithmFooter/AlgorithmFooter.types.ts
@@ -1,0 +1,5 @@
+export interface IAlgorithmFooter {
+  handleInitButton: () => void;
+  handleExcute: () => void;
+  handleSubmit: () => void;
+}
