+++ idearly/src/pages/AlgorithmSolvingPage/components/AlgorithmResult/AlgorithmResult.styles.tsx
@@ -2,6 +2,7 @@ import styled from "@emotion/styled";
 
 export const AlgorithmResultContainer = styled.div`
   width: 100%;
+  border-top: 1px solid #cccccc;
   border-bottom: 1px solid #cccccc;
   height: 18vh;
 `;
