+++ idearly/src/pages/AlgorithmSolvingPage/components/index.ts
@@ -5,3 +5,4 @@ export * from "./AlgorithmEditor/AlgorithmEditor";
 export * from "./AlgorithmProblem/AlgorithmProblem";
 export * from "./AlgorithmProHeader/AlgorithmProHeader";
 export * from "./AlgorithmResult/AlgorithmResult";
+export * from "./AlgorithmFooter/AlgorithmFooter";
